Refactor calculatedExpression to be a context
